fix(rag): address remaining PR #30 review — background indexing, semaphore, session_id required, float confidence, type fixes
- persist_node: MongoDB insert first, then fire-and-forget Qdrant indexing via create_task
- persist_batch: add Semaphore(5) to bound concurrency
- run_pipeline: dispatch RAG persistence as background tasks, gather at end
- run_layer/run_pipeline: make session_id required (no default), update all callers
- thinking_service: extract _persist_nodes_background with date/market fallbacks
- search_nodes: fix asyncio deadlock (module-level executor, direct arg passing)
- search_nodes: node_type schema changed to list[str] | None
- min_confidence: int → float across search.py, search_nodes.py, spec
- Tests updated for background indexing (asyncio.sleep(0) yields)

+++ backend/src/rag/search.py
@@ -32,7 +32,7 @@ class NodeSearchService:
         *,
         node_type: list[str] | None = None,
         sector: str | None = None,
-        min_confidence: int | None = None,
+        min_confidence: float | None = None,
         date_from: str | None = None,
         date_to: str | None = None,
         market: str | None = None,
